feat: implement order matching and trade management
- Add match.js for handling order matching and trade ID generation.
- Introduce parts.js for parsing keys related to trades.
- Create proof.js to verify fund proof wallet balances at match time.
- Implement schema.js for normalizing and validating order structures.
- Develop soul.js to generate unique identifiers for market windows.
- Establish Trade.js class for managing trade lifecycle and state.
- Add comprehensive tests in Trade.test.js to validate trade functionalities.
- Implement various trade actions: create, confirm, deposit, deliver, dispute, refund, and release.
- Enhance helper functions for resolving trade roles and managing trade records.
- Introduce time.js utility for managing time-related calculations.
