+++ sandbox/Lock.test.js
@@ -1,9 +1,9 @@
-import Test from "../Test.js"
-import { sha256 } from "../Utils/crypto.js"
-import { Lock } from "../Lock.js"
-import { devplatform } from "../Platform.js"
-import zen, { initZEN } from "../ZEN.js"
-import { HDNodeWallet, getBytes } from "../Ethers.js"
+import Test from "../src/core/Test.js"
+import { sha256 } from "../src/core/Utils/crypto.js"
+import { Lock } from "./Lock.js"
+import { devplatform } from "../src/core/Platform.js"
+import zen, { initZEN } from "../src/core/ZEN.js"
+import { HDNodeWallet, getBytes } from "../src/core/Ethers.js"
 
 const _SEA = zen
 await initZEN()
