+++ sandbox/Lock/index.js
@@ -1,4 +1,4 @@
-import { sha256 } from "../Utils/crypto.js"
+import { sha256 } from "../../src/core/Utils/crypto.js"
 
 // Domain separator ":TL:" or ":CL:" isolates index space per lock type
 // Prevents on-chain correlation between TL and CL of the same trade
