+++ src/core/Order/soul.js
@@ -1,25 +1,28 @@
-import zen from "../ZEN.js"
+import { now } from "/core/Utils.js"
+import ZEN from "/core/ZEN.js"
 
-// Returns the Pen soul string for one market window:
-// params = { baseId, side, candle }
-// key    = <timestamp>:<pub>:<nonce>
-//
-// The soul identity changes when baseId/side/candle change because they are encoded
-// in SEA.pen({ params }). Validation remains generic and reusable:
-//   seg 0 — timestamp whose derived candle must equal the soul candle
-//   seg 1 — full writer pub and must equal R[5]
-//   seg 2 — nonce (PoW uniqueness)
-export function soul({ baseId = this?.base?.id, side = this?.side, candle = Math.floor(Date.now() / 300000) } = {}) {
-    if (!baseId || !side) throw new Error("invalidInput")
-    const stamp = { tonum: { seg: { sep: ":", idx: 0, of: { reg: 0 } } } }
-    const writer = { seg: { sep: ":", idx: 1, of: { reg: 0 } } }
-    return zen.pen({
+// Construct and return order
+// The soul must be globally computable by everyone so that orders are discoverable.
+// It must not be unique for each order, but rather represent a "group" of orders that can be easily found and filtered by others.
+export function soul({ base, quote, side, candle } = {}) {
+    // The expected soul format:
+    // ZEN.pen() hex of: base asset (game item id), quote asset (currency id), order side,  candle number
+    base = base?.id || this?.base?.id
+    quote = quote?.id || this?.quote?.id
+    candle = candle || now() // default to current time candle
+    side = side || this?.side
+    if (!base || !quote || !side) throw new Error("invalidInput")
+    const key = { reg: 0 }
+    const sep = ":"
+    const timestamp = { tonum: { seg: { sep, idx: 0, of: key } } }
+    const writer = { seg: { sep, idx: 1, of: key } }
+    return ZEN.pen({
         key: {
             and: [
                 {
                     let: {
                         bind: 0,
-                        def: { divu: [stamp, 300000] },
+                        def: { divu: [timestamp, 300000] },
                         body: {
                             and: [
                                 { gte: [{ reg: 128 }, candle] },
@@ -29,12 +32,12 @@ export function soul({ baseId = this?.base?.id, side = this?.side, candle = Math
                     }
                 },
                 { eq: [writer, { reg: 5 }] },
-                { seg: { sep: ":", idx: 2, of: { reg: 0 }, match: { length: [1, 64] } } }
+                { seg: { sep, idx: 2, of: key, match: { length: [1, 64] } } }
             ]
         },
         val:  { type: "string" },
         sign: true,
         pow:  { field: 0, difficulty: 3 },
-        params: { baseId, side, candle }
+        params: { base, side }
     })
 }
